Login Cleanup of returned JSON from Google
Added some constants so that we only need to change Consatnts.cs and CustomUrlSchemeInterceptorActivity.cs to add the real Google API key. Also got  the returned JSON into an object that we can manipulte to easily get all the user's information.
Still don't know how to automatically close native web browser that is opened for authentication.